DOCUFY-DOCKER/flask-docker/Docufy/requirements.txt
#absl-py==0.9.0
absl-py>=1.0.0
astor==0.8.1
bson==0.5.8
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.0.2
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
#grpcio==1.28.1
gunicorn==19.9.0
#h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.14.1
#Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
#numpy==1.18.4
numpy>=1.21.6
oauthlib==3.1.0
opt-einsum==3.2.1
#Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.8.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
#scipy==1.4.1
scipy>=1.8.1
six==1.14.0
#tensorboard==2.0.2
#tensorflow==1.14.0
#tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==0.15.5
wrapt==1.12.1
tensorflow==2.10.1
grpcio>=1.47.0
h5py>=3.7.0
Pillow>=9.0.0

